+++ changelog.md
@@ -3,6 +3,8 @@
 - Add `Hist1d/2d.copy`
 - Add `Hist2d.from_image`
 - Add tutorial
+- `CoordinateSystem` and `CoordinateSystemArray` can now be initialized with
+  three vectors.
 
 # v5.4.1
 - Expose `for_pcolormesh_from_txt` and `for_pcolormesh_from_root`